Ot156 44 endpoint de eliminar usuario (#17)
* add controller users

* add services user

* add method delete
