+++ controllers/loginControllers.js
@@ -1,4 +1,5 @@
 const { login } = require('../services/loginServices');
+const { decodeAuthToken } = require('../services/jwtServices');
 
 const loginUser = async (req, res, next) => {
   try {
@@ -15,6 +16,16 @@ const loginUser = async (req, res, next) => {
   }
 };
 
+const getUserData = async (req, res, next) => {
+  try {
+    const user = decodeAuthToken(req.headers.authorization);
+    res.send(user);
+  } catch (err) {
+    next(err);
+  }
+};
+
 module.exports = {
   loginUser,
+  getUserData,
 };
