+++ routes/login.js
@@ -1,7 +1,6 @@
 const express = require('express');
 const validatorLogin = require('../middlewares/loginMiddlewares');
-const { getUserData } = require('../controllers/userControllers');
-const { loginUser } = require('../controllers/loginControllers');
+const { loginUser, getUserData } = require('../controllers/loginControllers');
 
 const router = express.Router();
