+++ routes/users.js
@@ -1,4 +1,5 @@
 const express = require('express');
+const { deleteUser } = require('../controllers/usersControllers');
 
 const router = express.Router();
 
@@ -6,5 +7,7 @@ const router = express.Router();
 router.get('/', function (req, res, next) {
   res.send('respond with a resource');
 });
+
+router.delete('/:id', deleteUser);
 
 module.exports = router;
