--- services/userServices.js
@@ -1,9 +1,0 @@
-const { User } = require('../models');
-
-const findEmail = async (email) => {
-  return await User.findOne({
-    where: { email },
-  });
-};
-
-module.exports = { findEmail };
