+++ services/usersServices.js
@@ -13,7 +13,24 @@ const getById = async (id) => {
   });
 };
 
+const findEmail = async (email) => {
+  return await User.findOne({
+    where: { email },
+  });
+};
+
+const remove = async (id) => {
+  return await User.update(
+    { deletedAt: new Date() },
+    {
+      where: { id },
+    }
+  );
+};
+
 module.exports = {
   getAllForCamps,
   getById,
+  findEmail,
+  remove,
 };
